+++ bin/Server.rb
@@ -1,10 +1,10 @@
 #!/usr/bin/env ruby
 
-require_relative '../lib/memcached/server/Server.rb'
+require 'ruby_memcached'
 
 puts('Creating server...')
 host = ARGV[0]
 port = ARGV[1]
-server = Memcached::Server.new(host, port)
+server = RubyMemcached::Server.new(host, port)
 puts('Starting server...')
 server.start()
