+++ lib/ruby_memcached.rb
@@ -1,0 +1,2 @@
+require_relative './ruby_memcached/Memcached.rb'
+require_relative './ruby_memcached/server/Server.rb'
