+++ lib/ruby_memcached/Responses.rb
@@ -1,0 +1,12 @@
+require 'ruby-enum'
+
+module RubyMemcached
+    class Memcached::Responses
+        include Ruby::Enum
+    
+        define :stored, 'STORED'
+        define :not_stored, 'NOT_STORED'
+        define :exists, 'EXISTS'
+        define :not_found, 'NOT_FOUND'
+    end
+end
