+++ lib/ruby_memcached/server/Errors.rb
@@ -1,0 +1,7 @@
+module RubyMemcached
+    class Server::Errors
+        ERROR = "ERROR\r\n"
+        CLIENT_ERROR = "CLIENT_ERROR %s\r\n"
+        SERVER_ERROR = "SERVER_ERROR %s\r\n"
+    end
+end
