--- lib/simple-memcached.rb
@@ -1,5 +1,0 @@
-require_relative './simple-memcached/Memcached.rb'
-require_relative './simple-memcached/Responses.rb'
-require_relative './simple-memcached/Item.rb'
-require_relative './simple-memcached/server/Server.rb'
-
