--- lib/simple-memcached/Responses.rb
@@ -1,10 +1,0 @@
-require 'ruby-enum'
-
-class Memcached::Responses
-    include Ruby::Enum
-
-    define :stored, 'STORED'
-    define :not_stored, 'NOT_STORED'
-    define :exists, 'EXISTS'
-    define :not_found, 'NOT_FOUND'
-end
