--- lib/simple-memcached/server/static/Errors.rb
@@ -1,5 +1,0 @@
-class Errors
-    ERROR = "ERROR\r\n"
-    CLIENT_ERROR = "CLIENT_ERROR %s\r\n"
-    SERVER_ERROR = "SERVER_ERROR %s\r\n"
-end
